July 7, 2021 9:56 AM

+++ Green Quests/Quest 3/Automaton.h
@@ -35,7 +35,7 @@ class Automaton {
         static size_t pow_2(size_t n) {return 1 << n;}
         static size_t translate_n_bits_starting_at(const vector<int>& bits, size_t pos, size_t n);
         size_t binary_to_decimal(const vector<int>& bits);
-        static vector<int> decimal_to_binary(size_t n);
+        static vector<int> decimal_to_binary(int n);
 
         friend class Tests;
 };
